Refactor codebase: Split code into smaller, more manageable modules for improved readability and maintainability.
